% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRmapping_functions.R
\name{analyze_benchmark}
\alias{analyze_benchmark}
\title{Analyze benchmark}
\usage{
analyze_benchmark(in_bm, in_measure, inp_resdir = NULL)
}
\arguments{
\item{in_bm}{\link[mlr3]{BenchmarkResult} or file name of one stored on disk.}

\item{in_measure}{\link[mlr3]{MeasureClassif} or \link[mlr3]{MeasureRegr} to use in model comparison}

\item{inp_resid}{(character) path to results directory where in_bm is stored}
}
\value{
list containing:
*a ggplot showing sensitivity, specificity, misclassification for different probability thresholds
*a ggplot boxplot showing the difference in performance measure between models
*a data.table with, for each model, the probability threshold value at which
the performance measure is maximized (and the corresponding performance measure value).
}
\description{
Evaluate the performance of multiple machine learning models.
}
\details{
this function is useful to optimize the flow intermittence probability
threshold above which to classify reaches as non-perennial (based on a specific
performance measure of interest).
}
