% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRmapping_functions.R
\name{analyzemerge_gaugeir}
\alias{analyzemerge_gaugeir}
\title{Quality assurance/quality checking of streamflow time series}
\usage{
analyzemerge_gaugeir(
  in_GRDCgaugestats,
  in_GSIMgaugestats,
  yearthresh,
  in_gaugep,
  inp_resdir,
  plotseries = FALSE
)
}
\arguments{
\item{in_GRDCgaugestats}{list of data.tables of intermittency statistics, output from
\code{\link{comp_GRDCdurfreq}} applied to all GRDC gauging stations.}

\item{in_GSIMgaugestats}{list of data.tables of intermittency statistics, output from
\code{\link{comp_GSIMdurfreq}} applied to all GSIM gauging stations.}

\item{yearthresh}{(integer) minimum year from which to analyze discharge record.}

\item{in_gaugep}{\link[sf]{sf} object of gauging stations.}

\item{inp_resdir}{(character) path to the results directory in which to create folder and write output plots}

\item{plotseries}{(logical) whether to create plots of streamflow time series for gauging stations
that are non-perennial only in winter or within 3 km from the coast.}
}
\value{
list of two data.tables.
\itemize{
\item A data.table listed as 'data' which contains the collated outputs from \code{\link{comp_GRDCdurfreq}}
and\code{\link{comp_GSIMdurfreq}} for the stations that were not excluded through this QA/QC process.
\item A data.table listed as flags, listing all stations excluded from further analysis and the reason their exclusion.
}
}
\description{
Record of quality-checking procedure which aimed to ensure the validity of
zero-flow readings and the flow intermittence class assigned to each gauge
(i.e., perennial or non-perennial).
}
\details{
For details on the QA/QC procedure, see Supplementary Information at
\link{https://www.nature.com/articles/s41586-021-03565-5}. The reason for
station exclusion is also presented in an interactive online map at
\link{https://messamat.github.io/globalIRmap/} in the Data and methods/Reference streamflow gauging stations tab.
}
