% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRmapping_functions.R
\name{bin_dt}
\alias{bin_dt}
\title{Bin data.table}
\usage{
bin_dt(
  in_dt,
  binvar,
  binfunc,
  binarg,
  bintrans = NULL,
  ndigits = 2,
  na.rm = FALSE,
  valuevar = NULL
)
}
\arguments{
\item{in_dt}{\link[data.table]{data.table} to bin.}

\item{binvar}{(character) column that will be used to define bins.}

\item{binfunc}{(character) binning approach. One of 'manual', 'equal_length', 'equal_freq'.}

\item{binarg}{(numeric) binning argument, depends on binning approach (\code{binfunc}).}

\item{bintrans}{(character or numeric) transformation of \code{binvar}, default is NULL.}

\item{ndigits}{(integer) number of decimals to keep for displaying formatted bin limits}

\item{na.rm}{(logical) whether to include NAs.}

\item{valuevar}{(character) if na.rm = FALSE, column to use to detect NAs and remove records.}
}
\value{
input \link[data.table]{data.table} with four new columns:
\itemize{
\item bin - bin number (1 being the lowest value bin)
\item bin_lmin - bin lower limit
\item bin_lmax - bin higher limit
\item bin_lformat - formatted character of bin limits
(format: \code{round(bin_lmin, ndigits) - round(bin_lmax, ndigits))
}
}
\description{
Bins a data.table over a numeric column.
}
\details{
inspired from \href{https://github.com/rsquaredacademy/rbin}{rbin package}.
Differences include that it concentrates all binning approaches within a single
function and works on a data.table.

binfunc: \cr
\itemize{
\item 'manual' - bin continuous data manually. \code{binarg} sets the inner bin limits,
such that the final table will have \code{length(binarg) + 1} bins. The lower end of the
first bin is automatically set to be the minimum value in \code{binvar} and the upper end of
the last bin is set to be the maximum value in \code{binvar}

\item 'equal_length' - Bin continuous data such that each bin has the same \code{binvar} interval length.
If \code{bintrans} is not \code{NULL}, then interval length is computed on transformed scale.
\code{binarg} (automatically rounded to the nearest integer) sets the number of bins.

\item 'equal_freq' - Bin continuous data such that each bin has the same number of records.
\code{binarg} (automatically rounded to the nearest integer) sets the number of bins.
}

bintrans: can either be 'log' (for natural log) or a numeric exponent to transform
according to x^bintrans.
}
\seealso{
for examples, see applications in \code{\link{bin_misclass}},
\code{\link{eval_watergap}}, \code{\link{tabulate_globalsummary}},
\code{\link{formathistab}}, \code{\link{compare_us}}
}
