% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRmapping_functions.R
\name{bin_misclass}
\alias{bin_misclass}
\title{Bin misclassification metrics}
\usage{
bin_misclass(
  in_predictions = NULL,
  in_resampleresult = NULL,
  in_gaugestats = NULL,
  binvar,
  binfunc,
  binarg,
  interthresh = 0.5,
  spatial_rsp = NULL,
  rspcol = NULL
)
}
\arguments{
\item{in_predictions}{a data.table of predictions for a set of CV repetitions.}

\item{in_gaugestats}{data.table with additional columns to use in binning if
not included as predictor variables in model.}

\item{binvar}{(character) column that will be used to define bins.}

\item{binfunc}{(character) binning approach. One of 'manual', 'equal_length', 'equal_freq'. See \link{bin_dt}.}

\item{binarg}{(numeric) binning argument, depends on binning approach. See \link{bin_dt}.}

\item{interthresh}{(numeric) flow intermittence probability threshold above which
to classify records as non-perennial}

\item{spatial_rsp}{whether to use results from spatial (TRUE) or non-spatial (FALSE)
cross-validation (only used if \code{in_rftuned} is provided, not \code{in_predictions}).}

\item{rspcol}{column to use that contains predicted probabilities (only used if
\code{in_predictions} is provided, not \code{in_rftuned}).}

\item{in_rftuned}{Output from \link{selecttrain_rf}; list containing inner and
outer resampling results + task.}
}
\value{
data.table with performance statistics by subset.
}
\description{
Get misclassification error, sensitivity, specificity, precision, and other
classification performance statistics for different subsets (i.e., bins) of
the dataset.
}
\details{
This function was used for internal assessment of model performance.
Right now, the options to change binvar and binfunc are not working
(need some adjustment); only manual binning based on long-term mean annual discharge ('dis_m3_pyr') is possible.
Update instances of 'dis_m3_pyr' for changing.
}
