% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRmapping_functions.R
\name{bm_msrtab}
\alias{bm_msrtab}
\title{Benchmark result performance measures table}
\usage{
bm_msrtab(bmres, interthresh = NULL)
}
\arguments{
\item{bmres}{\link[mlr3]{BenchmarkResult} from either a binary classification
model or a regression model bounded between 0 and 1.}

\item{interthresh}{(numeric) between 0 and 1 (inclusive), threshold to compute
sensitivity, specificity, and balanced accuracy. If not provided, specificity
and sensitivity will be provided for a 0.5 threshold and the highest BACC
for thresholds ranging from 0.45 to 0.55 will be provided.}
}
\value{
\link[data.table]{data.table} of model hyperparameters and
measure performance values for each \link[mlr3]{ResampleResult}.
}
\description{
Get hyperparameters and performance measures for every \link[mlr3]{ResampleResult}
in a \link[mlr3]{BenchmarkResult} of classification models.
}
\details{
performance measures include Binary brier score (bbrier),
Area Under the ROC Curve (AUC), sensitivity, specificity, and Balanced
classification ACCuracy (BACC).
}
\seealso{
\code{\link{bm_paramstime}}, \code{\link{rsmp_bbrier}},
\code{\link{rsmp_auc}}, \code{\link{rsmp_sen}}, \code{\link{rsmp_spe}},
\code{\link{rsmp_bacc}}
}
