% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRmapping_functions.R
\name{bm_paramstime}
\alias{bm_paramstime}
\title{Benchmark result parameters and time}
\usage{
bm_paramstime(bmres)
}
\arguments{
\item{bmres}{\link[mlr3]{BenchmarkResult} from either a binary classification
model or a regression model bounded between 0 and 1.}
}
\value{
\link[data.table]{data.table} of parameter values for each resample results
}
\description{
Get the parameters, their values or the range of tested values through tuning +
the running time for each resample result in a \code{\link[mlr3]{BenchmarkResult}}.
\cr
This is a utility function.
}
