% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRmapping_functions.R
\name{change_oversamp}
\alias{change_oversamp}
\title{Change oversampling ratio}
\usage{
change_oversamp(in_oversamprf, in_task)
}
\arguments{
\item{in_oversamprf}{\link[mlr3]{Learner} whose oversample.ratio parameter to update.}

\item{in_task}{\link[mlr3]{Task} based on which to determine minority class oversampling ratio.}
}
\value{
input \link[mlr3]{Learner} (\code{in_oversamprf}) with updated oversampling ratio.
}
\description{
Change the class oversampling ratio of an existing random forest \mlr3[mlr3]{Learner} object.
}
\details{
In this study, this function was created to adjust the oversampling ratio after updating
the criterion for classifying rivers and streams as non-perennial (see \link{change_taskbackend}).
The ratio of perennial to non-perennial gauges increases a lot when only considering watercourses
that cease to flow at least 30 days per year as non-perennial. This function enables one to keep
all settings and hyperparameters intact for a given Learner while updating that ratio.
}
