% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRmapping_functions.R
\name{change_tasktarget}
\alias{change_tasktarget}
\title{Change task back end}
\usage{
change_tasktarget(in_task, in_gaugestats, newmdurthresh)
}
\arguments{
\item{in_task}{\link[mlr3]{Task} whose target variable to update.}

\item{in_gaugestats}{data.table of task back-end content (i.e., data to update task with),
ontaining the column 'mDur_o1800' and all hydro-environmental predictor variables.}

\item{newmdurthresh}{(numeric) criterion in terms of mean annual number of zero-flow days over which to classify a river or stream as non-perennial (e.g., 30)}
}
\value{
\link[mlr3]{Task} with updated target variable.
}
\description{
Change target variable for task. This is implemented in this model to adjust
the criterion used to classify rivers and streams as non-perennial. Changing
the mean annual number of zero-flow day threshold (in the case of this study from 1 to 30 days).
}
