% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRmapping_functions.R
\name{checkGRDCzeroes}
\alias{checkGRDCzeroes}
\title{Check zero-flow values in GRDC discharge record}
\usage{
checkGRDCzeroes(
  GRDCstatsdt,
  in_GRDC_NO,
  period = 15,
  yearthresh,
  maxgap,
  in_scales = "free_x",
  labelvals
)
}
\arguments{
\item{GRDCstatsdt}{formatted data.table including a "path" column to access
GRDC standard daily discharge record text file for the gauging station of interest.}

\item{in_GRDC_NO}{GRDC unique identifier of the station to be investigated.}

\item{period}{(integer) number of days on each side of zero-flow values to subset.}

\item{yearthresh}{(integer) minimum year from which to analyze discharge record}

\item{maxgap}{(integer) threshold number of missing daily records to consider a calendar year unfit for analysis.}

\item{in_scales}{(character) should panel scales be fixed ("fixed", the default),
free ("free"), or free in one dimension ("free_x", "free_y")?}

\item{labelvals}{(logical) whether to label individual discharge values}
}
\value{
\link[data.table]{data.table} subset of GRDCstatsdt, including only
records for zero-flow days +- period
}
\description{
Plot and subset a given period (in days) on each side of every zero-flow
period in record for a given GRDC station. The output is a plot with a panel
showing discharge for each zero-flow period.
}
