% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRmapping_functions.R
\name{combine_bm}
\alias{combine_bm}
\title{Combine benchmarking results}
\usage{
combine_bm(in_resampleresults, write_qs = NULL, inp_resdir = NULL)
}
\arguments{
\item{in_resampleresults}{one or a list of \link[mlr3]{ResampleResult} objects to convert to \link[mlr3]{BenchmarkResult}}

\item{write_qs}{whether to write out serialized (qs) file}

\item{inp_resdir}{path to directory where to write out qs file}
}
\value{
file name of qs file
}
\description{
Combine a list of \link[mlr3]{ResampleResult} into a single \link[mlr3]{BenchmarkResult}
object and save the object in serialized form on disk.
}
\details{
qs file is writen to \code{inp_resdir/combine_bmYYMMDDHHSS.qs}
}
