% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRmapping_functions.R
\name{comp_derivedvar}
\alias{comp_derivedvar}
\title{Compute derived variables}
\source{
Linke, S., Lehner, B., Dallaire, C. O., Ariwi, J., Grill, G., Anand,
M., ... & Tan, F. (2019). Global hydro-environmental sub-basin and river
reach characteristics at high spatial resolution. Scientific Data, 6(1),
1-15.
}
\usage{
comp_derivedvar(in_dt, copy = FALSE)
}
\arguments{
\item{copy}{(logical) whether to make a copy of \code{dt} (if TRUE) or to
modify it in place (if FALSE)}

\item{dt}{data.table which contains rows as records and HydroATLAS
environmental variables as columns.}
}
\description{
Format and compute derived a set of environmental variables from input
dataset that contains
\href{https://www.hydrosheds.org/page/hydroatlas}{HydroATLAS} variables for
the purpose of predicting intermittency.
}
\details{
This function only formats and compute derived variables deemed
relevant for the intermittency analysis. \cr
\cr
Steps include: \cr
\enumerate{
\item Convert some -9999 that should be 0 after checking them
\item Count the number of -9999 values per column
\item Convert the rest of the -9999 values to NA
\item Compute a set of derived variables based on existing variables in
HydroATLAS
(e.g. \code{pre_mm_cvar= fifelse(pre_mm_cmx==0, 0, pre_mm_cmn/pre_mm_cmx)})
\item Correct scaling from RiverATLAS
(e.g. Degree of regulation is out of 10,000 in HydroATLAS)
}\preformatted{}
}
