% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRmapping_functions.R
\name{comp_durfreq}
\alias{comp_durfreq}
\title{Compute intermittency statistics for a gauging station}
\usage{
comp_durfreq(path, maxgap, monthsel = NULL)
}
\arguments{
\item{path}{file path to a GRDC-formatted streamflow time series table}

\item{maxgap}{maximum number of days with missing data beyond which a year is
not used in the computation of statistics}

\item{monthsel}{selected months to compute the statistics over}
}
\value{
One row data.table with the following columns: \cr
#For a given gauge discharge file (given by path), format and compute
}
\description{
Determine general characteristics of the whole time series and of the subset
of years that have less than a given threshold of missing data as well as
intermittency statistics. The intermittency statistics can be computed for a
subset of months of the year (e.g. only winter months)
}
