% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRmapping_functions.R
\name{comp_irstats}
\alias{comp_irstats}
\title{Compute flow intermittence statistics}
\usage{
comp_irstats(tabyearly, maxgap, mdurthresh, yearthresh, windowsize, fullwindow)
}
\arguments{
\item{tabyearly}{data.table}

\item{maxgap}{(integer) maximum number of days with missing data beyond which a year is
not used in the computation of statistics.}

\item{mdurthresh}{(numeric) threshold of mean annual number of zero-flow days beyond
which to classify gauge as intermittent.}

\item{yearthresh}{(integer) minimum year from which to analyze discharge record.}

\item{windowsize}{(integer) window size to check for zero-flow days.}

\item{fullwindow}{(logical) whether years for which the window is truncated
(e.g., beginning and end of time series) are taken in account in moving window analysis.}
}
\value{
link[data.table]{data.table} with 1 row and 10 columns:
\itemize{
\item firstYear_kept: the first year in the time series with a number of missing daily values <= \code{maxgap}
(i.e., the first year in the subset of the time series kept for further analysis)
\item lastYear_kept: the last year in the time series with a number of missing daily values <= \code{maxgap}
(i.e., the first year in the subset of the time series kept for further analysis)
\item totalYears_kept: total number of years in the time series with a number of missing daily values <= \code{maxgap}
(i.e., the number of years in the subset of the time series kept for further analysis)
\item totaldays: total number of days with data in years with a number of missing daily values <= \code{maxgap}
(i.e., total number of days of data kept for further analysis)
\item integerperc: proportion of daily values which are in integer format.
Only including years with <= maxgap missing daily discharge values.
\item sumDur: total number of daily zero-flow values.
Only including years with <= maxgap missing daily discharge values.
\item mDur: mean annual number of daily zero-flow values.
Only including years with <= maxgap missing daily discharge values.
\item mFreq: mean annual frequency of zero flow events. A zero-flow event is defined
as one or more consecutive days with a recorded daily discharge of zero.
Only including years with <= maxgap missing daily discharge values.
\item intermittent: binary flow intermittence class. 1: non-perennial (if mDur >= 1, i.e.,
if gauging station recorded zero-flow for at least one day per year on average); 0: perennial.
\item movinginter: whether there is at least one zero-flow day in every \code{windowsize}-year (e.g., 20-year) moving window across the record.
}
}
\description{
Format and compute a set of summary statistics based on a yearly streamflow
gauging station time series. Used in \code{\link{comp_GRDCdurfreq}}.
}
