% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRmapping_functions.R
\name{comp_ymean}
\alias{comp_ymean}
\title{Compute annual mean}
\usage{
comp_ymean(in_dt, fieldex, mstart, outcol)
}
\arguments{
\item{in_dt}{data table, contains columns of monthly values.}

\item{fieldex}{character, example column name containing monthly value.
It must include the month number in \%m format (two digits e.g. 02 for February)}

\item{mstart}{integer; position of the first digit of the month in the column
of monthly values.}

\item{outcol}{character; name of new column to be created containing monthly mean}
}
\value{
Modifies in_dt in place. Add column of weighted mean across months
(based on number of days in months)
}
\description{
Compute annual mean of a variable based on columns of monthly means.
}
