% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRmapping_functions.R
\name{compare_au}
\alias{compare_au}
\title{Compare model estimates for Australia}
\usage{
compare_au(inp_resdir, in_rivpred, predcol, binarg)
}
\arguments{
\item{inp_resdir}{(character) full path to directory containing formatted
network and basins data for Australia.}

\item{in_rivpred}{data.table of model predictions for global river network. Here, output from \link{netpredformat}.}

\item{predcol}{(character) name of the column for the predicted probability
of flow intermittence generated by this study (Messager et al. 2021).}

\item{binarg}{(numeric) limits of the mean annual flow bins across which to compare estimates (and for plotting histogram). See \link{bin_dt}.}
}
\value{
list with two elements:
*"plot": histogram comparing estimated prevalence of flow intermittence and
total river length across discharge size classes.
*"data": summary comparison statistics.
}
\description{
Compare model estimates generated for Australia by this study to model
estimates in the Australian National Geofabric by producing comparative histograms
and statistics.
}
\details{
this function is used to produce Extended Data Figure 5 in Main Text from Messager et al. (2021).
Only rivers and stream segments with a drainage area >= 10 km2 are included in the comparison.
}
