% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRmapping_functions.R
\name{compare_fr}
\alias{compare_fr}
\title{Compare model estimates for France}
\usage{
compare_fr(inp_frdir, in_rivpred, predcol, binarg, mincutoff)
}
\arguments{
\item{inp_frdir}{(character) full path to directory containing formatted
network and basins for France.}

\item{in_rivpred}{data.table of model predictions for global river network. Here, output from \link{netpredformat}.}

\item{predcol}{(character) name of the column for the predicted probability
of flow intermittence generated by this study (Messager et al. 2021).}

\item{binarg}{(numeric) limits of the mean annual flow bins across which to compare estimates (and for plotting histogram). See \link{bin_dt}.}

\item{mincutoff}{(numeric) smallest river reaches to include in comparison, in terms of long-term mean annual flow (m3/s; e.g., 0.1)}
}
\value{
list with two elements:
*"plot": histogram comparing estimated prevalence of flow intermittence and
total river length across discharge size classes.
*"data": summary comparison statistics.
}
\description{
Compare model estimates generated for mainland France by this study to model
estimates generated by Snelder et al. (2013) by producing comparative histograms
and statistics.
}
\details{
this function is used to produce Extended Data Figure 5 in Main Text from Messager et al. (2021).
}
