% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRmapping_functions.R
\name{compare_us}
\alias{compare_us}
\title{Compare model estimates for the U.S.}
\usage{
compare_us(inp_usresdir, inp_usdatdir, in_rivpred, predcol, binarg, mincutoff)
}
\arguments{
\item{inp_usresdir}{(character) full path to directory containing formatted
network and basins data for the U.S.
@param inp_usdatdir (character) full path to directory containing raw data
for the U.S.}

\item{in_rivpred}{data.table of model predictions for global river network. Here, output from \link{netpredformat}.}

\item{predcol}{(character) name of the column for the predicted probability
of flow intermittence generated by this study (Messager et al. 2021).}

\item{binarg}{(numeric) limits of the mean annual flow bins across which to compare estimates (and for plotting histogram). See \link{bin_dt}.}

\item{mincutoff}{(numeric) smallest river reaches to include in comparison, in terms of long-term mean annual flow (m3/s; e.g., 0.1)}
}
\value{
list with two elements:
*"plot": histogram comparing estimated prevalence of flow intermittence and
total river length across discharge size classes.
*"data_all": summary comparison statistics assuming that line segments in the
NHDPlus classified as "Artificial path" are all perennial.
*"data_noartificial": summary comparison statistics excluding line segments
classified as "Artificial path" are all perennial.
}
\description{
Compare model estimates generated for the contiguous U.S. by this study to the
prevalence of intermittence as depicted in the National Hydrography Dataset
Plus (NHDPlus) medium-resolution (100k) by producing comparative histograms
and statistics.
}
\details{
this function is used to produce Extended Data Figure 5 in Main Text from Messager et al. (2021).
}
