% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRmapping_functions.R
\name{compute_confustats}
\alias{compute_confustats}
\title{Compute confusion statistics}
\usage{
compute_confustats(in_gselpreds, ndigits = 2)
}
\arguments{
\item{in_gselpreds}{data.table with a row for each record and two columns, 'truth' and 'response'.}

\item{ndigits}{number of decimals to include in misclassification, sensitivity,
specificity, and precision metrics.}
}
\value{
vector of statistics
}
\description{
Compute standard set of statistics from a confusion matrix based on classification
results (including misclassification rate, sensitivity, specificity, precision,
number of perennial and non-perennial stations correctly and incorrectly classified,
total number of stations, and the predicted prevalence of flow intermittence compared
to the observed prevalence).
}
\details{
Formatted statistics for Extended Data Table 2 in Messager et al. (2021).
}
