% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRmapping_functions.R
\name{compute_IRpop}
\alias{compute_IRpop}
\title{Compute population living nearest to an intermittent river}
\usage{
compute_IRpop(in_rivpred, inp_linkpop, valuevar)
}
\arguments{
\item{in_rivpred}{data.table of model predictions for global river network. output from \link{netpredformat}.}

\item{inp_linkpop}{absolute path to table identifying number of people living
nearest to each river reach in the global river network.}

\item{valuevar}{name of the column in \code{in_rivpred} by which to compute
percentage of the population (i.e., column name that contains the predicted
binary flow intermittence class for each river reach)}
}
\value{
two-row data.table with the number and percentage of people living nearest
to non-perennial vs. perennial river reaches.
}
\description{
Compute total population and percentage of the world's population which live
nearest to a river or stream reach predicted to be non-perennial.
}
