% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRmapping_functions.R
\name{convert_clastoregrtask}
\alias{convert_clastoregrtask}
\title{Convert classification task to regression task}
\usage{
convert_clastoregrtask(in_task, in_id, oversample = FALSE)
}
\arguments{
\item{in_task}{\link[mlr3]{TaskClassif} or \link[mlr3spatiotempcv]{TaskClassifST}}

\item{in_id}{the id of the output task}

\item{oversample}{(logical) whether to oversample minority class beforehand
(this is not exactly equivalent to a PipeOp) but oversample PipeOp is not yet
available for regression tasks. If TRUE, will oversample so that classes are
equally represented in task.}
}
\value{
a \link[mlr3]{TaskRegr}
}
\description{
Convert a binary \link[mlr3]{TaskClassif} or
\link[mlr3spatiotempcv]{TaskClassifST}
to a \link[mlr3]{TaskRegr}, optionally oversampling the minority class first.
}
\examples{
\dontrun{
in_dt <- data.table(intermittent=c(rep(0, 300), rep(1, 300)))
task = mlr3::TaskClassif$new(id = "task_classif", backend = in_dt, target ='intermittent')
convert_clastoregrtask(task, id="task_regr")
}

}
\seealso{
\code{\link{create_tasks}}
}
