% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRmapping_functions.R
\name{create_baselearners}
\alias{create_baselearners}
\title{Create learners}
\usage{
create_baselearners(in_task)
}
\arguments{
\item{in_task}{\link[mlr3]{Task} of type |link[mlr3spatiotempcv]{TaskClassif},
|link[mlr3spatiotempcv]{TaskClassifST}, or |link[mlr3]{TaskRegr}.}
}
\value{
This function will return different types of learners depending on the type of task input. \cr
If a |link[mlr3spatiotempcv]{TaskClassif} or |link[mlr3spatiotempcv]{TaskClassifST} is provided,
then a list is returned with six learners, a classif.ranger (standard probability random forest learner) and
classif.cforest (probability conditional inference forest learner), each in three formats for class
imbalance correction: 1. without any class imbalance correction (e.g., lrn_ranger),
2. with random oversampling of minority class (e.g., 'lrn_ranger_overp), and 3. with class weighting (e.g., 'lrn_ranger_weight'). \cr

If a |link[mlr3]{TaskRegr} is provided, then a single learner is returned of type 'regr.ranger'
with maximally selected rank statistics (maxstat) as splitting rule. \cr

For all learners, 800 trees are grown.
}
\description{
Create machine learning \link[mlr3]{Learner} list for random forest model.
}
\details{
Learners are methods to train and predict a model for a Task and provide
meta-information about the learners, such as the hyperparameters you can set.
For more info on mlr3 Learners, see \link{https://mlr3book.mlr-org.com/learners.html}.
}
