% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRmapping_functions.R
\name{create_tasks}
\alias{create_tasks}
\title{Create tasks}
\usage{
create_tasks(
  in_gaugestats,
  in_predvars,
  id_suffix = NULL,
  include_discharge = TRUE
)
}
\arguments{
\item{in_gaugestats}{data.table of formatted gauging station summary statistics and hydro-environmental attributes. Output from \link{format_gaugestats}.}

\item{in_predvars}{data.table of candidate predictor variable codes and names. Output from \link{selectformat_predvars}.}

\item{id_suffix}{(character) suffix to add after task names}

\item{include_discharge}{(logical) whether to include predictor variables of long-term discharge modeled from WaterGAP v2.2.}
}
\value{
list of three \link[mlr3]{Task}:
\itemize{
\item(a spatiotemporal classification task) |link[mlr3spatiotempcv]{TaskClassifST};
for classification random forest with spatiotemporal CV (standard probability and probability CIF)
\item(a standard regression tasks) |link[mlr3]{TaskRegr};
for regression random forest (MAXSTAT)
\item(a regression task where the minority class has already been oversampled)
|link[mlr3]{TaskRegr}; for regression random forest (MAXSTAT).
Implemented because class oversampling pipelines are not available for regression tasks.
}
}
\description{
Create machine learning \link[mlr3]{Task} list for random forest models.
}
\details{
Tasks are objects that contain the (usually tabular) data and
additional meta-data to define a machine learning problem. The meta-data is,
for example, the name of the target variable for supervised machine learning
problems, or the type of the dataset (e.g. a spatial or survival).
This information is used for specific operations that can be performed on a task.
For more info on tasks, see \link{https://mlr3book.mlr-org.com/tasks.html}.
}
