% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRmapping_functions.R
\name{def_filestructure}
\alias{def_filestructure}
\title{Define file structure}
\usage{
def_filestructure()
}
\value{
Named list with paths to directories and files as well as paths for
analysis output
}
\description{
Defines file structure for running RF prediction of global intermittency
}
\details{
all paths are relative to file in which project/package is contained. \cr
Project must be contained in ~/project_name/src \cr
Data must be contained in ~/project_name/data \cr
A directory called ~/project_name/results must exist to write outputs
}
