% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRmapping_functions.R
\name{diny}
\alias{diny}
\title{Number of days in the year}
\usage{
diny(year)
}
\arguments{
\item{year}{Numeric or integer vector of the year (format: 4-digit year, \%Y).}
}
\value{
Number of days in that year.
}
\description{
Computes the number of days in a given year, taking in account leap years
}
\examples{
diny(1999)
diny(2000)
diny(2004)
diny(2100)
diny(1600)

}
