% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRmapping_functions.R
\name{eval_watergap}
\alias{eval_watergap}
\title{Compute modeled discharge statistics for GRDC gauging stations}
\usage{
eval_watergap(in_qstats, in_selgauges, binarg)
}
\arguments{
\item{in_qstats}{data.table of discharge statistics for GRDC streamflow gauging stations.
Each row corresponds to a single station. Output from \link{comp_GRDCqstats}.}

\item{in_selgauges}{data.table or data.frame of gauging stations to analyze.}

\item{binarg}{discharge bin limits to divide performance statistics asssessment by
(to create size classes based on long-term mean annual flow in m3/s).}
}
\value{
data.table of performance statistics by streamflow size class for
mean annual flow and Q90.
}
\description{
Compute performance statistics (regression R2 and sMAPE) for WaterGAP v2.2 discharge
predictions of long-term mean discharge and Q90 against observed discharge
at GRDC gauging stations with ≥20 years of streamflow data.
}
\details{
gauging stations to analyze \code{in_selgauges}  are further subsetted
to keep only those with at least 20 years of daily discharge data.

This function performs a log-log regression across all gauges and non-log regressions
for each size class to compute R2.

This function was used to produce Table S1 in the Supplementary Information of Messager et al. 2021
}
