% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRmapping_functions.R
\name{extend_globalsummary_clz}
\alias{extend_globalsummary_clz}
\title{Extend global summary}
\usage{
extend_globalsummary_clz(in_IRESextra, in_globaltable, inp_riveratlas_legends)
}
\arguments{
\item{in_IRESextra}{estimate of prevalence of flow intermittence in global river
network for river reaches with 0.01 < MAF 0.099}

\item{in_globaltable}{summary data.table of the predicted prevalence of IRES
across discharge size classes and climate zones}

\item{inp_riveratlas_legends}{(character) full path to .xlsx table containing
metadata on river atlas attributes (meaning of category acronyms and codes e.g., for land cover)}
}
\value{
data.table
}
\description{
Add statistics for extrapolated discharge size class to table of summary statistics
on global prevalence
}
\details{
this function is used to produce Table 1 in Main Text, Extended Data Figure 1c, and
Supplementary Data tables from Messager et al. (2021).
}
