% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRmapping_functions.R
\name{extract_impperf_nestedrf}
\alias{extract_impperf_nestedrf}
\title{Extract variable importance and performance from a trained RF learner}
\usage{
extract_impperf_nestedrf(
  in_rflearner,
  in_task,
  imp = TRUE,
  perf = TRUE,
  pvalue = TRUE,
  pvalue_permutn = 100
)
}
\arguments{
\item{in_rflearner}{An \link[mlr3]{AutoTuner} or a trained
\link[mlr3]{Learner}.}

\item{imp}{(logical) Whether to compute variable importance.}

\item{perf}{(logical) Whether to compute the performance measure.}

\item{pvalue}{(logical) Whether to compute p-values.}

\item{pvalue_permutn}{(integer) number of permutations to use in p-value
calculations}
}
\value{
A data.table with, when all logicals are TRUE, the following columns.
\describe{
\item{varnames} - predictor variable name
\item{imp} - variable importance value
\item{pvalue} - variable importance p-value
\item{[perf]} - performance value (same for all rows), name changes
}
}
\description{
Computes the performance, variable importance and associated p-value from
either a trained \link[mlr3]{AutoTuner} or a trained \link[mlr3]{Learner}.
}
\details{
Accepts both a trained \link[mlr3]{AutoTuner} or a trained \
link[mlr3]{Learner} of class \link[mlr3learners]{mlr_learners_classif.ranger}
or \link[mlr3learners.partykit]{mlr_learners_classif.cforest}. \cr

Also accept learners of class \link[mlr3pipelines]{GraphLearner} \cr

If \code{p-value==TRUE} and \link[mlr3learners]{mlr_learners_classif.ranger},
then compute p-values with \link[ranger]{importance_pvalues} with Altmann
permutation method using \code{pvalue_permutn} permutations. \cr

If \link[mlr3learners.partykit]{mlr_learners_classif.cforest} is provided,
\code{p-value} is ignored for now.
}
\section{documentation to-do}{

Can add an example down the line, add source.
}

\seealso{
\code{\link{weighted_vimportance_nestedrf}}
}
