% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRmapping_functions.R
\name{extrapolate_IRES}
\alias{extrapolate_IRES}
\title{Extrapolate prevalence of flow intermittence}
\usage{
extrapolate_IRES(
  in_rivpred,
  in_extranet,
  min_cutoff = 0.1,
  interactive = F,
  valuevar = "predbasic800cat",
  grouping_var = "PFAF_IDclz"
)
}
\arguments{
\item{in_rivpred}{output from \link{netpredformat}.}

\item{in_extranet}{output from \link{extrapolate_networklength}.}

\item{min_cutoff}{(numeric) minimum discharge to include in model training.}

\item{interactive}{(logical) whether to print results and make plots for user
to interactively evaluate results and troubleshoot.}

\item{valuevar}{(character) name of the column containing the beta-distributed
response variable to extrapolate (based on mean annual flow as the indepedent variable).}

\item{grouping_var}{variable with which to subset the dataset into groups. A separate
model is trained on each group.}
}
\value{
list containing a data.table and two plots.
The data.table contains an estimate
}
\description{
Train models and use them to extrapolate the global prevalence of intermittent
rivers and ephemeral streams (as a percentage of network length).
}
\details{
The prevalence of IRES  was independently extrapolated for a total
of 465 spatial sub-units representing all occurring intersections of
62 river basin regions (BasinATLAS level 2 subdivisions) and 18 climate zones
(Global Environmental Stratification). For each basin–climate sub-unit, after
extrapolating the empirical cumulative distribution of total stream length
(of all reaches with MAF ≥ 0.1 m3 s−1) down to 0.01 m3/s MAF, we extrapolated
the prevalence of flow intermittence (in percentage of stream length) down to
0.01 m3/s MAF. wefitted a GAM for beta-distributed data—that is,
with a (0, 1) range—to the prevalence of intermittence in each logarithmic
MAF size bin of the sub-unit.
}
