% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRmapping_functions.R
\name{extrapolate_networklength}
\alias{extrapolate_networklength}
\title{Extrapolate river network length}
\usage{
extrapolate_networklength(
  inp_riveratlas,
  min_cutoff = 0.1,
  dispred = seq(0.01, 0.09, 0.01),
  interactive = T,
  grouping_var = "PFAF_IDclz"
)
}
\arguments{
\item{inp_riveratlas}{full path to RiverATLAS (v1.0) attribute table.}

\item{min_cutoff}{(numeric) minimum discharge to include in model training.}

\item{dispred}{(numeric vector) discharge values for which to produce model predictions.}

\item{interactive}{(logical) whether to print results and make plots for user
to interactively evaluate results and troubleshoot.}

\item{grouping_var}{variable with which to subset the dataset into groups. A separate
model is trained on each group.}
}
\value{
list containing a data.table and two plots.
The data.table contains an estimate stream length for each discharge size class
and climate-basin sub-unit. cumL_pred is the cumulative length of all rivers
and streams with MAF > dispred. cumL_cutoffref is the cumulative river length
at MAF == min_cutoff. cumL_predextra = cumL_pred - cumL_cutoffref.
}
\description{
Train models and use them to extrapolate the global length of rivers.
}
\details{
The prevalence of IRES  was independently extrapolated for a total
of 465 spatial sub-units representing all occurring intersections of
62 river basin regions (BasinATLAS level 2 subdivisions) and 18 climate zones
(Global Environmental Stratification). For each basin–climate sub-unit, we
first extrapolated the empirical cumulative distribution of total stream length
(of all reaches with MAF ≥ 0.1 m3 s−1) down to 0.01 m3 s−1 MAF using a generalized
additive model (GAM). We excluded reaches larger than the 95th percentile of
MAF (that is, the largest rivers) within the sub-unit from model fitting to avoid common
discontinuities at the high end of the empirical distribution
that can affect the low end of the power-law-like trendline.
}
