% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRmapping_functions.R
\name{format_modelcompdat}
\alias{format_modelcompdat}
\title{Format model computational metadata}
\usage{
format_modelcompdat(bmres, typecomp = c("classif1", "regr1", "classif2"))
}
\arguments{
\item{bmres}{\link[mlr3]{BenchmarkResult}}

\item{typecomp}{(character) one of 'classif1', 'regr1', 'classif2'. Type of model.}
}
\value{
\link[data.table]{data.table} of BenchmarkResult with formatted model names
}
\description{
Format model Learner names for each learner in a \link[mlr3]{BenchmarkResult}. \cr
Utility function.
}
