% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRmapping_functions.R
\name{formathistab}
\alias{formathistab}
\title{Format histogram table}
\usage{
formathistab(
  in_dt,
  castvar,
  valuevar,
  valuevarsub,
  weightvar = 1,
  binfunc,
  binarg,
  binlabels,
  datname = NULL
)
}
\arguments{
\item{in_dt}{\link[data.table]{data.table} to bin.}

\item{castvar}{(character) olumn that will be used to define bins —
the equivalent of \code{binvar} in \code{bin_dt}.}

\item{valuevar}{(character) variable to summarize (e.g. intermittency class).}

\item{valuevarsub}{(character) value of \code{valuevar} to summarize (e.g. '1' for intermittent rivers)}

\item{weightvar}{(character) variable to weigh proportion by. (e.g. river reach length).}

\item{binfunc}{(character) binning approach. One of 'manual', 'equal_length', 'equal_freq'.}

\item{binarg}{(numeric) binning argument, depends on binning approach (\code{binfunc}).}

\item{binlabels}{(character vector) formatted labels for bins.}

\item{datname}{(character) optional, adds a column to output table describing what the data describes (e.g. France)}
}
\value{
\link[data.table]{data.table} with five columns:
\itemize{
\item bin - bin number
\item perc - percentage of records (or of \code{weightvar}) that meet \code{valuevar == valuevarsub}.
\item binsumlength - number of records (if default \code{weightvar}) or sum of \code{weightvar} by bin (e.g. total river length).
\item dat - \code{datname} argument
\item binformat - label for each bin provided through \code{binlabels} argument
}
}
\description{
Creates a binary frequency histogram table by computing the proportion of records
for which a selected column has a given value
(in this study, the percentage length of rivers that are deemed intermittent)
after binning the table by a continuous variable (e.g. river discharge).
}
\seealso{
used in \code{\link{compare_fr}}, \code{\link{compare_us}},
and \code{\link{compare_au}}
}
