% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRmapping_functions.R
\name{formatscales}
\alias{formatscales}
\title{Format plot scales}
\usage{
formatscales(in_df, varstoplot)
}
\arguments{
\item{in_df}{data.frame with all records for environmental variables.
Used to determine the appropriate range of values for each variable. In this case,
a data.table of the river network hydro-environmental attributes.}

\item{varstopplot}{vector of variable names that will be plots and for which
to return a list of scales}
}
\value{
list of x and y scale objects + cartesian coordinates for ggplot
}
\description{
Utility function to format plot scales for density distribution plots of environmental variables.
}
