% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRmapping_functions.R
\name{fread_cols}
\alias{fread_cols}
\title{fread columns}
\usage{
fread_cols(file_name, cols_tokeep)
}
\arguments{
\item{file_name}{path of the table to be read}

\item{cols_tokeep}{character vector, names of columns to read}
}
\value{
a data.table with the \code{cols_tokeep} that are found in the table
}
\description{
Fast data.table-based reading of a subset of columns from a table
}
\examples{
fread_cols(iris, c('Sepal.Length', 'Sepal.Width')

}
