% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRmapping_functions.R
\name{get_basemapswintri}
\alias{get_basemapswintri}
\title{Get and project basemaps}
\usage{
get_basemapswintri()
}
\value{
list containing basemaps and graticule in Winkel Tripel projection
}
\description{
Get and project basemaps into Winkel Tripel for subsequent mapping of gauging stations.
}
