% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRmapping_functions.R
\name{get_outerrsmp}
\alias{get_outerrsmp}
\title{Get outer ResampleResult}
\usage{
get_outerrsmp(in_rftuned, spatial_rsp = FALSE)
}
\arguments{
\item{in_rftuned}{(list or \link[mlr3]{ResampleResult}) Output from
\code{\link{selecttrain_rf}}.}

\item{spatial_rsp}{(logical) Whether to return the results from the
spatial resampling strategy if there is one.}
}
\value{
a \link[mlr3]{ResampleResult}.
}
\description{
Extract the outer resampling \link[mlr3]{ResampleResult} from the output
from \code{\link{selecttrain_rf}} with the option to grab the output from
the spatial resampling strategy or not (when there are multiple resampling
strategies).
}
\details{
If \code{in_rftuned} is already a \link[mlr3]{ResampleResult}, then simply
return it.
}
\section{Warning}{

This will only return the \link[mlr3]{ResampleResult} from a single resampling
strategy.
}

