% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRmapping_functions.R
\name{get_oversamp_ratio}
\alias{get_oversamp_ratio}
\title{Get oversample ratio}
\usage{
get_oversamp_ratio(in_task)
}
\arguments{
\item{in_task}{binary \link[mlr3]{TaskClassif }}
}
\value{
named list with following items:
\describe{
\item{minoclass} - Value of minority class (e.g. '1' for intermittent rivers)
\item{ratio} - numeric ratio between number of items in majority and minority
class
}
}
\description{
Identify minority class and compute ratio between the number of observations
in the majority and the minority classes of a binary
\link[mlr3]{TaskClassif }.
}
\examples{
\dontrun{
in_dt <- data.table(intermittent=c(rep(0, 300), rep(1, 300)))
task = mlr3::TaskClassif$new(id = "in_dt", backend = in_dt, target ='intermittent')
get_oversamp_ratio(task)
}

}
