% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRmapping_functions.R
\name{get_USdata}
\alias{get_USdata}
\title{Compute intermittency statistics for US gauging stations}
\usage{
get_USdata(in_ids, maxyear, verbose = F)
}
\arguments{
\item{in_ids}{(character vector) unique USGS identifiers for gauging stations (e.g., "05540130")}

\item{maxyear}{(integer) maximum year to include in computation of intermittency statistics}

\item{verbose}{whether to print the \code{in_ids}}
}
\value{
data.table with time series record size and intermittency statistics computed by
\code{\link{comp_irstats}}
}
\description{
Download original daily data from the USGS National Water Information System
for U.S. stations and compute intermittency statistics.
}
\details{
This function is implemented in \code{\link{analyzemerge_gaugeir}} to compare
intermittency statistics obtained directly from USGS data compared to those obtained from
GSIM. This was implemented because it appears that GSIM rounded discharge data to two decimals
when converting cubic feet to cubic meters per second. Therefore, any discharge alue < 0.005 was considered zero.
}
