% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRmapping_functions.R
\name{ggalluvium_gaugecount}
\alias{ggalluvium_gaugecount}
\title{Alluvium plot of gauge count over time}
\usage{
ggalluvium_gaugecount(dtformat, alluvvar)
}
\arguments{
\item{dtformat}{formatted data.table (generated in \code{\link{analyzemerge_gaugeir}}).}

\item{alluvvar}{grouping ID to identify unique records}
}
\value{
plot
}
\description{
Utility function: create an \href{https://corybrunson.github.io/ggalluvial/}{alluvial plot}
of the number of streamflow gauging stations over time.
}
