% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRmapping_functions.R
\name{ggcompare}
\alias{ggcompare}
\title{Comparison ggplot histogram}
\usage{
ggcompare(datmerge, binarg, insetx = 0.4, insety = 0.8)
}
\arguments{
\item{datmerge}{\link[data.table]{data.table}}

\item{binarg}{labels for inner bin limits, see \link{bin_dt}}

\item{insetx}{(numeric) proportion of main plot width to occupy with inset (between 0 and 1)}

\item{insety}{(numeric) proportion of main plot height to occupy with inset (between 0 and 1)}
}
\value{
histogram \link[ggplot2]{ggplot} object
}
\description{
utility function: creates a ggplot histogram showing side-by-side bars
to compare two sets of data for the same bins. Based on \link[data.table]{data.table}
containing formatted histogram data.
}
\details{
\code{datmerge} \link[data.table]{data.table} must contain all the columns
returned by \link{formathistab}.
}
\seealso{
used in \code{\link{compare_fr}}, \code{\link{compare_us}},
and \code{\link{compare_au}}
}
