% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRmapping_functions.R
\name{ggenvhist}
\alias{ggenvhist}
\title{Plot of environmental histogram}
\usage{
ggenvhist(vartoplot, in_gaugedt, in_rivdt, in_predvars, scalesenvhist)
}
\arguments{
\item{vartoplot}{(column) variable for which to produce a density plot.}

\item{in_gaugedt}{data.table of gauging stations' environmental attributes.}

\item{in_rivdt}{data.table of global river network's environmental attributes}

\item{in_predvars}{data.table of predictor variable codes, names and attributes.
Output from \link{selectformat_predvars}.}

\item{scalesenvhist}{list of scale objects to format plot. From \link{formatscales}.}
}
\value{
ggplot with two density distributions of the environmental variable,
one for the gauging stations and one for the global river network.
}
\description{
Utility function to create an individual density plot of the distribution of a
given environmental variables across gauges and the whole global river network.
}
