% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRmapping_functions.R
\name{gggauges}
\alias{gggauges}
\title{Map river gauging stations}
\usage{
gggauges(in_gaugepred, in_basemaps, binarg, binvar)
}
\arguments{
\item{in_gaugepred}{data.table of predictions for gauging stations. Output from \link{bind_gaugepreds}.}

\item{in_basemaps}{list of basemaps \link\link{sf} objects.}

\item{binarg}{(vector) inner bin breaks to plot histogram for gauging stations.}

\item{binvar}{(character) column to plot histogram of gauging stations with}
}
\value{
ggplot (patchwork)
}
\description{
Make two maps of streamflow gauging stations used in model training and testing.
One map of perennial stations and the other one of non-perennial stations.
}
\details{
The maps include a histogram of streamflow record duration in each map.
binarg c(30, 60, 100) is supplied to show four bins in the histogram (10-30, 30-60, 60-100, >100).

This function was used to produce the data underlying Figure 3 of the Main Text of Messager et al. 2021.
see \link{test_joincount} for more details on the procedure for testing spatial autocorrelation.
}
