% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRmapping_functions.R
\name{ggpartialdep}
\alias{ggpartialdep}
\title{Plot of partial dependence}
\usage{
ggpartialdep(
  in_rftuned,
  in_predvars,
  colnums,
  ngrid,
  nodupli = T,
  nvariate = 2,
  parallel = T,
  spatial_rsp = FALSE
)
}
\arguments{
\item{in_rftuned}{Output from \link{selecttrain_rf}; list containing inner and
outer resampling results + task.}

\item{in_predvars}{data.table of predictor variable codes, names and attributes.
Output from \link{selectformat_predvars}.}

\item{colnums}{number of variables to include, in decreasing order of variable importance.}

\item{ngrid}{number of predictor variable values to check model's marginally predicted value for.}

\item{nodupli}{whether to include variable types only once (i.e. if minimum discharge is in, not including mean discharge)}

\item{nvariate}{(1 or 2) whether to analyze univariate (1) or bivariate (2)
partial dependence.}

\item{parallel}{(boolean) whether to compute function in parrallel
(as it can be computationally intensive).}

\item{spatial_rsp}{(boolean) whether to use outputs from spatial (TRUE) or
non-spatial (FALSE) cross-validation.}
}
\value{
pages of gridded plots with each page containing 9 plots
}
\description{
Produce line or raster ggplots of univariate or bivariate partial dependence.
i.e., estimates of the marginal relationship between predictor variables and
the model’s predictions (probability of intermittence) by holding the rest of
the predictors at their respective mean values. Bivariate plots show the co-linearity
in response between two predictors.
}
\details{
this function is used to produce Figure S5 in the Supplementary Information for Messager et al. 2021.
This function was initially developed for bivariate plots but only univariate plots
were produced for the final manuscript so there may be obsolete snippets left.
}
