% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRmapping_functions.R
\name{ggrivers}
\alias{ggrivers}
\title{Plot river gauging stations basemaps}
\usage{
ggrivers(in_basemaps)
}
\arguments{
\item{in_basemaps}{list of basemaps \link\link{sf} objects.}
}
\value{
ggplot
}
\description{
Utility function to format plot of basic basemap to show the distribution of river gauging stations.
}
\details{
does not include rivers in basemaps. Too long to draw, too messy/
}
