% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRmapping_functions.R
\name{ggvimp}
\alias{ggvimp}
\title{Plot of variable importance}
\usage{
ggvimp(in_rftuned, in_predvars, varnum = 10, spatial_rsp = FALSE)
}
\arguments{
\item{in_rftuned}{Output from \link{selecttrain_rf}; list containing inner and
outer resampling results + task.}

\item{in_predvars}{data.table of predictor variable codes, names and attributes.
Output from \link{selectformat_predvars}.}

\item{varnum}{number of variables whose variable importance to plot.}

\item{spatial_rsp}{(boolean) whether to use variable importance values derived from
spatial (TRUE) or non-spatial (FALSE) cross-validation.}
}
\value{
ggplot
}
\description{
Produce a bar ggplot of variable importance for random forest model.
}
\details{
this function is used to produce Figure 2 in the Main Text of Messager et al. 2021.
Additional formatting is needed to exactly match Fig. 2.
}
