% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRmapping_functions.R
\name{label_manualbins}
\alias{label_manualbins}
\title{Label manual bins}
\usage{
label_manualbins(binarg, minval)
}
\arguments{
\item{binarg}{(character vector) Arguments for bin_dt manual}

\item{minval}{(numeric) value to set for lower limit of first bin.}
}
\value{
vector of labels
}
\description{
Utility function: label bin limits for \code{\link{formathistab}}.
}
