% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRmapping_functions.R
\name{layout_ggenvhist}
\alias{layout_ggenvhist}
\title{Layout plots of environmental histograms}
\usage{
layout_ggenvhist(in_rivernetwork, in_gaugepred, in_predvars)
}
\arguments{
\item{in_rivernetwork}{data.table of global river network's environmental attributes. Here, output from \link{rformat_network}.}

\item{in_gaugepred}{selected gauging stations' environmental attributes. Here, output from \link{write_gaugepreds}.}

\item{in_predvars}{data.table of predictor variable codes, names and attributes.
Output from \link{selectformat_predvars}.}
}
\value{
ggplots with two density distributions of the environmental variable,
one for the gauging stations and one for the global river network.
}
\description{
Run plotting functions across predictor variables and arrange plots
}
\details{
function used to produce Extended Data Fig. 8  c-p in Messager et al. 2021.
}
