% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRmapping_functions.R
\name{make_gaugepreds}
\alias{make_gaugepreds}
\title{Make gauge predictions}
\usage{
make_gaugepreds(
  in_rftuned,
  in_gaugestats,
  in_res_spcv = NULL,
  in_predvars,
  interthresh,
  simple = FALSE
)
}
\arguments{
\item{in_rftuned}{trained random forest model, output from \link{ selecttrain_rf}}

\item{in_gaugestats}{data.table of formatted gauging station summary statistics and hydro-environmental attributes. Output from \link{format_gaugestats}.}

\item{in_res_spcv}{\link[mlr3]{ResampleResult} from spatial cross-validation, output from \link{dynamic_resamplebm}.}

\item{in_predvars}{data.table of predictor variable codes, names and attributes. See \link{selectformat_predvars}.}

\item{interthresh}{(numeric) between 0 and 1 (inclusive), probability threshold
above which to classify gauging stations as non-perennial (e.g., 0.50).}

\item{simple}{(logical) if TRUE, return only final model predictions, if FALSE return cross-validation predictions}
}
\value{
\code{in_gaugestats} data.table with one new column (IRpredprob_full, if simple == T)
or five new columns (if simple == F). \cr
\itemize{
\item IRpredprob_full: predicted probability that station is non-perennial based on final model training.
\item IRpredprob_CVsp: predicted probability that station is non-perennial based on spatial cross-validation
(i.e., prediction when fold whose station is a member was excluded from training)
\item IRpredprob_CVnosp: predicted probability that station is non-perennial based on spatial cross-validation
\item inter_class_u10_featsel_spfold orinter_class_o1_featsel_spfold: cross-validation fold membership (id)
\}
}
\description{
Make predictions for all gauging stations used in model training based on final trained model.
Join outputs of prediction and cross-validations to original gauging stations statistics table.
}
