% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRmapping_functions.R
\name{map_basinBACC}
\alias{map_basinBACC}
\title{Map prediction performance at the basin level}
\usage{
map_basinBACC(
  in_gaugepred,
  in_rivernetwork,
  inp_basin,
  outp_basinerror,
  spatial_rsp
)
}
\arguments{
\item{in_gaugepred}{data.table of predictions for gauging stations. Output from \link{bind_gaugepreds}.}

\item{in_rivernetwork}{river network with model predictions and other attributes. Output from \link{netpredformat}.}

\item{inp_basin}{(character) absolute path to feature class of HydroBASINS level 3 polygons}

\item{outp_basinerror}{(character) absolute path of output basin polygons with computed statistics}

\item{spatial_rsp}{(boolean) whether to use outputs from spatial (TRUE) or
non-spatial (FALSE) cross-validation.}
}
\value{
list containing a data.table and two plots.
The data.table contains an estimate
}
\description{
Compute the number of gauging stations, bias, residual spatial autocorrelation
and balanced classification accuracy for each HydroBASIN level 3.
}
\details{
This function was used to produce the data underlying Figure 3 of the Main Text of Messager et al. 2021.
see \link{test_joincount} for more details on the procedure for testing spatial autocorrelation.
}
