% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRmapping_functions.R
\name{netpredformat}
\alias{netpredformat}
\title{Format network predictions}
\usage{
netpredformat(outp_riveratlaspred, in_rivernetwork)
}
\arguments{
\item{outp_riveratlaspred}{full path to model predictions for river network.}

\item{in_rivernetwork}{data.table of formatted global river network with all
predictor hydro-environmental variables. Output from \link{rformat_network}.}
}
\value{
data.table with model predictions and other river network attributes, including
river reach length (LENGTH_KM), long-term mean annual discharge (dis_m3_pyr),
the percentage of the river reach length that intersects with a lake (INLAKEPERC),
the surface area of the river reach's full upstream drainage area (UPLAND_SKM), and
the climate zone where the river reach is located (clz_cl_cmj).
}
\description{
Read model predictions for network and add river reach and basin identifiers,
geographic information, and hydrological attributes.
}
\details{
the date is added to the output file name in the format \code{outp_riveratlaspred_YYYYMMDD}.
}
