% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRmapping_functions.R
\name{plot_coastalir}
\alias{plot_coastalir}
\title{Plot coastal intermittent rivers}
\usage{
plot_coastalir(
  in_gaugep = in_gaugep,
  dt = GRDCstatsdt,
  yearthresh,
  dbname = "grdc",
  inp_resdir = inp_resdir,
  plotseries = TRUE
)
}
\arguments{
\item{in_gaugep}{\link[sf]{sf} object of gauges with column called class99_19_rsp9_buf3k1
indicating whether the gauge is located within 3 km of marine waters.}

\item{dt}{formatted data.table (generated in \code{\link{analyzemerge_gaugeir}}).}

\item{yearthresh}{(integer) minimum year from which to plot discharge record}

\item{dbname}{(character) 'GRDC' or 'GSIM". Will determine whether to use \code{\link{plotGRDCtimeseries}}
or \code{\link{plotGSIMtimeseries}}}

\item{inp_resdir}{(character) path to directory where to save plots. A directory will be
created called " 'winterir_rawplots_o(yearthresh)_(date)'}

\item{plotseries}{(logical) whether to generate plots}
}
\value{
subset of input dt only including only intermittent rivers within 3 km of a coastline
}
\description{
Utility function: plot discharge time series for GRDC or GSIM gauging stations
that have been deemed to be near marine waters.
}
\details{
so-called marine stations were defined as those within 3 km of a coastline.
}
