% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRmapping_functions.R
\name{plot_GRDCflags}
\alias{plot_GRDCflags}
\title{Plot all GRDC time series with data quality flags}
\usage{
plot_GRDCflags(
  in_GRDCgaugestats,
  yearthresh,
  inp_resdir,
  maxgap,
  showmissing = FALSE
)
}
\arguments{
\item{in_GRDCgaugestats}{data.table (or list of data.tables) with time series
and intermittency statistics for all GRDC gauging stations.}

\item{yearthresh}{(integer) minimum year from which to plot discharge record.}

\item{inp_resdir}{(character) path to the results directory in which to create folder and write output plots}

\item{maxgap}{(integer) threshold number of missing daily records to consider a calendar year unfit for analysis.}

\item{showmissing}{(logical) whether to show records in years with number of missing daily records beyond \code{maxgap}.}
}
\value{
nothing (empty data.table)
}
\description{
Creates pngs of streamflow time series plots of daily discharge ({m^3}/s; with flags
for 0-flow values and potential outlier) for all GRDC gauges with at least
10 years of data, excluding years with more than 20 days of missing data.
}
\details{
the output graphs are written into two separate newly created directories in \code{inp_resdir} called
GRDCir_rawplots_\code{yearthresh}_\code{YYYYMMDD} and GRDCper_rawplots_\code{yearthresh}_\code{YYYYMMDD}
(e.g., GRDCir_rawplots_1800_20200512 and GRDCper_rawplots_1800_20200512). The first contains plots for non-perennial
gauging stations and the second contains plots for perennial gauging stations. \cr
\cr
Each plot is generated by \code{\link{plotGRDCtimeseries}} and shows
the time series of daily streamflow values for a station.
For the flagging criteria, see documentation for \code{\link{flagGRDCoutliers}}.
\itemize{
\item The y-axis is square-root transformed.
\item Individual points show daily discharge values (in {m^3}/s).
\item blue lines link daily values (which may result in unusual patterns due to missing years).
\item red points are zero-flow flow values.
\item green points are non-zero flow daily values statistically flagged as potential outliers .
\item black points are zero-flow values flagged as potential outliers.
}
}
