% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRmapping_functions.R
\name{plot_GSIM}
\alias{plot_GSIM}
\title{Plot all GSIM time series with data quality flags}
\usage{
plot_GSIM(in_GSIMgaugestats, yearthresh, inp_resdir, maxgap, showmissing)
}
\arguments{
\item{in_GSIMgaugestats}{data.table (or list of data.tables) with time series
and intermittency statistics for all GSIM gauging stations.}

\item{yearthresh}{(integer) minimum year from which to plot discharge record.}

\item{inp_resdir}{(character) path to the results directory in which to create folder and write output plots}

\item{maxgap}{(integer) threshold number of missing daily records to consider a calendar year unfit for analysis.}

\item{showmissing}{(logical) whether to show records in years with number of missing daily records beyond \code{maxgap}.}
}
\value{
nothing (empty data.table)
}
\description{
Creates pngs of streamflow time series plots of monthly discharge ({m^3}/s)
for all GSIM gauges with at least 10 years of data, excluding years with more
than 20 days of missing data.
}
\details{
the output graphs are written into two separate newly created directories in \code{inp_resdir} called
GSIMir_rawplots_\code{yearthresh}_\code{YYYYMMDD} and GSIMper_rawplots_\code{yearthresh}_\code{YYYYMMDD}
(e.g., GSIMir_rawplots_1800_20200512 and GSIMper_rawplots_1800_20200512). The first contains plots for non-perennial
gauging stations and the second contains plots for perennial gauging stations. \cr
\cr

Each plot is generated by \code{\link{plotGSIMtimeseries}}.
Daily streamflow records from GSIM stations are unavailable. Therefore,
the graph shows the following:
\itemize{
\item The y-axis is square-root transformed.
\item Blue points: mean monthly discharge
\item Light blue background shading: mean ± 2SD monthly discharge
\item Black points: minimum and maximum monthly discharge
\item Red points show minimum monthly discharge values equal to 0
\item Purple points show months for which all daily discharge values are equal to 0.
}
}
