% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRmapping_functions.R
\name{plot_winterir}
\alias{plot_winterir}
\title{Plot winter-only intermittent rivers}
\usage{
plot_winterir(dt, dbname, inp_resdir, yearthresh, plotseries = TRUE)
}
\arguments{
\item{dt}{formatted data.table (generated in \code{\link{analyzemerge_gaugeir}}).}

\item{dbname}{(character) 'GRDC' or 'GSIM". Will determine whether to use \code{\link{plotGRDCtimeseries}}
or \code{\link{plotGSIMtimeseries}}}

\item{inp_resdir}{(character) path to directory where to save plots. A directory will be
created called " 'winterir_rawplots_o(yearthresh)_(date)'}

\item{yearthresh}{(integer) minimum year from which to plot discharge record}

\item{plotseries}{(logical) whether to generate plots}
}
\value{
subset of input dt only including winter-only intermittent irvers
}
\description{
Utility function: plot discharge time series for GRDC or GSIM gauging stations
that have been deemed to monitor winter-only intermittent rivers.
}
\details{
“Winter-only” non-perennial gauging stations were defined as those
whose stream record contained less than one zero-flow day per year on average
during months with long-term mean air temperature over 10°C (averaged across
the local catchment immediately draining to the river reach, according to
WorldClim 2). In other words, “winter-only” non-perennial gauging stations
were those which would not have qualified as non-perennial according to our
criterion if only non-winter months were taken into account.
}
