% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRmapping_functions.R
\name{plotGRDCtimeseries}
\alias{plotGRDCtimeseries}
\title{Plot a GRDC time series}
\usage{
plotGRDCtimeseries(
  GRDCgaugestats_record,
  outpath = NULL,
  maxgap = 366,
  showmissing = FALSE
)
}
\arguments{
\item{GRDCgaugestats_record}{\link[data.table]{data.table} of formatted daily
discharge records for a single GRDC station. In this project, e.g. the output
from \code{\link{comp_GRDCdurfreq}}. Must contain at least five columns:
\code{GRDC_NO, dates, Original, flag_mathis, missingdays} \cr
Alternatively, a named list or vector with
a column called "path" towards a standard GRDC text file containing daily discharge
records for a single gauging station.}

\item{outpath}{(character) path for writing output png plot (default is no plotting).}

\item{maxgap}{(integer) threshold number of missing daily records to consider a calendar year unfit for analysis.}

\item{showmissing}{(logical) whether to show records in years with number of missing daily records beyond \code{maxgap}.}
}
\value{
plot
}
\description{
Creates a plot of daily discharge ({m^3}/s) for a GRDC gauging station,
with flags for 0-flow values and potential outliers.
Save plot to png if path is provided.
}
\details{
the output graphs show the time series of daily streamflow values
for the station. For the flagging criteria, see documentation for \code{\link{flagGRDCoutliers}}.
\itemize{
\item The y-axis is square-root transformed.
\item Individual points show daily discharge values (in {m^3}/s).
\item blue lines link daily values (which may result in unusual patterns due to missing years).
\item red points are zero-flow flow values.
\item green points are non-zero flow daily values statistically flagged as potential outliers .
\item black points are zero-flow values flagged as potential outliers.
}
}
