% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRmapping_functions.R
\name{plotGSIMtimeseries}
\alias{plotGSIMtimeseries}
\title{Plot a GSIM time series}
\usage{
plotGSIMtimeseries(
  GSIMgaugestats_record,
  outpath = NULL,
  maxgap = 366,
  showmissing = FALSE
)
}
\arguments{
\item{GSIMgaugestats_record}{a data containing structure  with
a column called "path" towards a standard montly GSIM text file containing.
In this project, e.g. the output from \code{\link{comp_GSIMdurfreq}}.}

\item{outpath}{(character) path for writing output png plot (default is no plotting).}

\item{maxgap}{(integer) threshold number of missing daily records to consider a calendar year unfit for analysis.}

\item{showmissing}{(logical) whether to show records in years with number of missing daily records beyond \code{maxgap}.}
}
\value{
plot
}
\description{
Creates a plot of monthly discharge ({m^3}/s) for a GSIM gauging stations
Save plot to png if path is provided.
}
\details{
Daily streamflow records from GSIM stations are unavailable. Therefore,
the graph shows the following:
\itemize{
\item The y-axis is square-root transformed.
\item Blue points: mean monthly discharge
\item Light blue background shading: mean ± 2SD monthly discharge
\item Black points: minimum and maximum monthly discharge
\item Red points show minimum monthly discharge values equal to 0
\item Purple points show months for which all daily discharge values are equal to 0.
}
}
