% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRmapping_functions.R
\name{qc_onde}
\alias{qc_onde}
\title{Quality check model estimates in France}
\usage{
qc_onde(
  inp_ondedatdir,
  inp_onderesdir,
  inp_riveratlas,
  in_rivpred,
  predcol,
  interthresh = 0.5,
  mincutoff = 0
)
}
\arguments{
\item{inp_ondedatdir}{(character) full path to directory containing raw
data for France.}

\item{inp_onderesdir}{(character) full path to directory containing formatted
data for France.}

\item{inp_riveratlas}{(character) full path to .csv. table containing attributes
for global river network}

\item{in_rivpred}{data.table of model predictions for global river network. Here, output from \link{netpredformat}.}

\item{predcol}{(character) name of the column for the predicted probability
of flow intermittence generated by this study (Messager et al. 2021).}

\item{interthresh}{(numeric) flow intermittence probability threshold above which
to classify records as non-perennial.}

\item{mincutoff}{(numeric) smallest river reaches to include in comparison, in terms of long-term mean annual flow (m3/s; e.g., 0.1)}
}
\value{
list with two elements:
*"plot": scatterplots analyzing Intermittence Prediction Residuals (IPR) as a function of
modeled mean annual flow, the position of the field observation along the RiverATLAS river reach,
the ratio of the drainage area of the watercourse at the location of the field
observation to that for the pourpoint of the corresponding RiverATLAS river reach,
the percentage of field observations that are either no-flow or dry,
the number of field observations at the site, and the population density in the entire upstream area
of the corresponding reach.
*"stats": summary comparison statistics.
}
\description{
Analyze performance of model predictions of flow intermittence against ONDE field
observations of flow state across mainland France.
}
\details{
this function is used to format data used in producing Extended Data Figure 6.
A side effect of this function is the writing of a .gpkg of river reaches for
which there are flow observations in France with quality-checking
attributes.
}
