% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRmapping_functions.R
\name{read_GSIMgauged_paths}
\alias{read_GSIMgauged_paths}
\title{Read file paths to gauge flow data}
\usage{
read_GSIMgauged_paths(inp_GSIMindicesdir, in_gaugep, timestep)
}
\arguments{
\item{in_gaugep}{table containing column named \code{GRDC_NO} with the
gauge IDs that will be used to generate file path.}
}
\value{
vector of paths to GRDC-formatted streamflow time series tables
}
\description{
Based on selection of gauges, create a list of paths to streamflow data
associated with gauges.
}
