% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRmapping_functions.R
\name{read_monthlydis}
\alias{read_monthlydis}
\title{Read network monthly discharge}
\usage{
read_monthlydis(in_path)
}
\arguments{
\item{in_filestructure}{named list containing path to RiverATLAS spatial data
with monthly discharge, named \code{in_monthlynetdischarge}.}
}
\value{
Data frame of monthly discharge for every reach in RiverATLAS
}
\description{
Import Naturalized mean monthly discharge from WaterGAP v2.2 downscaled to
the HydroSHEDS river network (12 monthly values for each river reach)
}
