% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRmapping_functions.R
\name{readformatGSIMsea}
\alias{readformatGSIMsea}
\title{Read and pre-format GSIM seasonal data}
\usage{
readformatGSIMsea(path)
}
\arguments{
\item{path}{(character) path to the text file of seasonal discharge data in
standard GSIM format (.mon file).}
}
\value{
\link[data.table]{data.table} of monthly indices, with additional attributes including
the estimated minimum number of zero-flow days "mDur_minmo"
}
\description{
Reads text file of seasonal discharge indices for a single GSIM station.
Creates columns for year, month, and season
Estimate the minimum possible number of monthly zero-flow days based on indices
}
\details{
the monthly minimum number of zero flow days n0 is estimated with
percentile statistics like:
if P90==0, then n0 = as.integer(floor(0.9\emph{n.available)),
if else P80==0, then n0 = as.integer(floor(0.8}n.available)),
if else P70 ...
...
else if MIN7 == 0, then n0 = 7
else if MIN == 0, then n0 = 1,
else n0 = 0

For more information on GSIM, see \href{https://essd.copernicus.org/articles/10/787/2018/}{Gudmundsson et al. (2018)}.
}
