% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRmapping_functions.R
\name{reset_tuning}
\alias{reset_tuning}
\title{Reset AutoTuner}
\usage{
reset_tuning(in_autotuner, in_task, in_lrnid = NULL)
}
\arguments{
\item{in_autotuner}{\link[mlr3tuning]{AutoTuner} (single or list)}

\item{in_task}{new \link[mlr3]{Task} to feed into the \link[mlr3tuning]{AutoTuner}.}

\item{in_lrnid}{ID for the \link[mlr3]{Learner} to select if a list of \link[mlr3tuning]{AutoTuner} was provided.}
}
\value{
\link[data.table]{data.table} of model hyperparameters and
measure performance values for each \link[mlr3]{ResampleResult}.
}
\description{
Create a new \link[mlr3tuning]{AutoTuner} based on the \link[mlr3]{Learner}
and other hyperparameters for an input AutoTuner, but replace with a new
\link[mlr3]{Task}.
}
\details{
performance measures include Binary brier score (bbrier),
Area Under the ROC Curve (AUC), sensitivity, specificity, and Balanced
classification ACCuracy (BACC).
}
\seealso{
\code{\link{bm_paramstime}}, \code{\link{rsmp_bbrier}},
\code{\link{rsmp_auc}}, \code{\link{rsmp_sen}}, \code{\link{rsmp_spe}},
\code{\link{rsmp_bacc}}
}
