% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRmapping_functions.R
\name{rformat_network}
\alias{rformat_network}
\title{Format river network}
\usage{
rformat_network(
  in_predvars,
  in_monthlydischarge = NULL,
  inp_riveratlasmeta,
  inp_riveratlas,
  inp_riveratlas2
)
}
\arguments{
\item{in_predvars}{data.table of predictor variable codes, names and attributes. See \link{selectformat_predvars}.}

\item{in_monthlydischarge}{optional data.table of long-term mean monthly
discharge values (from WaterGAP v2.2) for global river reaches.}

\item{inp_riveratlasmeta}{path to RiverATLAS attributes metadata table.}

\item{inp_riveratlas}{path to RiverATLAS (v1.0) attribute table (for the entire global river network).}

\item{inp_riveratlas2}{path to new attributes calculated for global river network following RiverATLAS methodology}
}
\value{
data.table of identifiers and hydro-environmental attribute for all reaches in the
RiverATLAS global river
}
\description{
Format river network, joining data, selecting variables, computing derived variables, etc.
}
