% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRmapping_functions.R
\name{rsmp_auc}
\alias{rsmp_auc}
\title{Resample AUC}
\source{
Hanley, J. A., & McNeil, B. J. (1982). The meaning and use of the
area under a receiver operating characteristic (ROC) curve. Radiology.
https://doi.org/10.1148/radiology.143.1.7063747
}
\usage{
rsmp_auc(bmres, rsmp_i)
}
\arguments{
\item{bmres}{\link[mlr3]{BenchmarkResult} from either a binary classification
model or a regression model bounded between 0 and 1.}

\item{rsmp_i}{(integer) index of the resample result to process.}
}
\value{
(numeric) auc
}
\description{
Compute the Area Under the ROC Curve or AUC (Hanley and McNiel, 1982) for binary
classification predictions from a single \link[mlr3]{ResampleResult} extracted
from an mlr3 \code{\link[mlr3]{BenchmarkResult}}. \cr
This is a utility function.
}
\details{
See the \href{https://mlr3.mlr-org.com/reference/mlr_measures_classif.auc.html}{mlr3 reference website}
for a definition of the AUC.
}
