% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRmapping_functions.R
\name{rsmp_bacc}
\alias{rsmp_bacc}
\title{Resample BACC}
\source{
Brodersen, K. H., Ong, C. S., Stephan, K. E., & Buhmann, J. M. (2010).
The balanced accuracy and its posterior distribution. Proceedings -
International Conference on Pattern Recognition, 3121–3124.
https://doi.org/10.1109/ICPR.2010.764
}
\usage{
rsmp_bacc(bmres, rsmp_i, threshold_class)
}
\arguments{
\item{bmres}{\link[mlr3]{BenchmarkResult} from either a binary classification
model or a regression model bounded between 0 and 1.}

\item{rsmp_i}{(integer) index of the resample result to process.}
}
\value{
(numeric) bacc
}
\description{
Compute the Balanced class ACCuracy (Brodersen, Ong, Stephan, & Buhmann, 2010)
from a single \link[mlr3]{ResampleResult} extracted from an mlr3
\code{\link[mlr3]{BenchmarkResult}}. \cr
This is a utility function.
}
\details{
See the \href{https://mlr3.mlr-org.com/reference/mlr_measures_classif.bacc.html}{mlr3 reference website}
for a definition of the BACC.
}
