% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRmapping_functions.R
\name{rsmp_bbrier}
\alias{rsmp_bbrier}
\title{Resample binary brier score}
\source{
Brier, G. W. (1950). Verification of forecasts expressed in terms of
probability. Monthly Weather Review, 78(1), 1\–3.
}
\usage{
rsmp_bbrier(bmres, rsmp_i)
}
\arguments{
\item{bmres}{\link[mlr3]{BenchmarkResult} from either a binary classification
model or a regression model bounded between 0 and 1.}

\item{rsmp_i}{(integer) index of the resample result to process.}
}
\value{
(numeric) binary brier score
}
\description{
Compute the binary brier score (Brier, 1950) for binary classification
predictions from a single \link[mlr3]{ResampleResult} extracted from an mlr3
\code{\link[mlr3]{BenchmarkResult}}. \cr
This is a utility function.
}
\details{
See the \href{https://mlr3.mlr-org.com/reference/mlr_measures_classif.bbrier.html}{mlr3 reference website}
for a definition of the Brier score for binary classification.
}
