% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRmapping_functions.R
\name{rsmp_pre}
\alias{rsmp_pre}
\title{Resample precision}
\usage{
rsmp_pre(bmres, rsmp_i, threshold_class)
}
\arguments{
\item{bmres}{\link[mlr3]{BenchmarkResult} from either a binary classification
model or a regression model bounded between 0 and 1.}

\item{rsmp_i}{(integer) index of the resample result to process.}
}
\value{
(numeric) precision
}
\description{
Compute the precision (proportion of stations classified as non-perennial that are actually IRES)
from a single \link[mlr3]{ResampleResult} extracted from an mlr3
\code{\link[mlr3]{BenchmarkResult}}. \cr
This is a utility function.
}
\details{
Precision is computed as the number of true positives (i.e. the
number of items correctly labelled as belonging to the positive class) divided
by the total number of elements labelled as belonging to the positive class.
\href{https://en.wikipedia.org/wiki/Precision_and_recall}{Wikipedia}
actually has a nice explanation for it.
}
