% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRmapping_functions.R
\name{rsmp_spe}
\alias{rsmp_spe}
\title{Resample specificity}
\usage{
rsmp_spe(bmres, rsmp_i, threshold_class)
}
\arguments{
\item{bmres}{\link[mlr3]{BenchmarkResult} from either a binary classification
model or a regression model bounded between 0 and 1.}

\item{rsmp_i}{(integer) index of the resample result to process.}
}
\value{
(numeric) specificity
}
\description{
Compute the specificity (proportion of correctly classified perennial stations)
from a single \link[mlr3]{ResampleResult} extracted from an mlr3
\code{\link[mlr3]{BenchmarkResult}}. \cr
This is a utility function.
}
\details{
Specificity measures the proportion of negatives that are correctly
identified. \href{https://en.wikipedia.org/wiki/Sensitivity_and_specificity}{Wikipedia}
actually has a nice explanation for it.
}
