% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRmapping_functions.R
\name{scientific_10}
\alias{scientific_10}
\title{Format number to nearest log10 bin in scientific format}
\usage{
scientific_10(x)
}
\arguments{
\item{x}{(numeric)}
}
\value{
(expression)
}
\description{
Format a number to nearest order of magnitude (log10 integer) and gives out a
scientific format expression.
}
