% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRmapping_functions.R
\name{select_features}
\alias{select_features}
\title{Select features}
\usage{
select_features(in_bm, in_lrnid, in_task, pcutoff, inp_resdir = NULL)
}
\arguments{
\item{in_bm}{\link[mlr3]{BenchmarkResult} containing the \link[mlr3]{ResampleResult}
for the learner of interest, based on which to select predictor variables. Can also be name of
serialized file (.qs) on disk (accessed through \link[qs]{qread}) containing BenchmarkResult object.}

\item{in_lrnid}{id of learner to extract from \link[mlr3]{BenchmarkResult} (e.g., "oversample.classif.ranger")}

\item{in_task}{\link[mlr3]{Task} containing predictor variables to subset}

\item{pcutoff}{variable p-value cut-off under which to select variable, see \link{extract_impperf_nestedrf} and
\link[ranger]{importance_pvalues} for more information on p-value calculation.}

\item{inp_resdir}{(character) path to where qs file is located (excluding the name of the qs file)}
}
\value{
list containing the original \code{in_task} and the updated task only included selected predictor variables.
}
\description{
Select a subset of model predictor variables (aka features) based on variable importance p-value.
}
