% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRmapping_functions.R
\name{selectformat_predvars}
\alias{selectformat_predvars}
\title{Select + format predictor variables}
\usage{
selectformat_predvars(inp_riveratlas_meta, in_gaugestats)
}
\arguments{
\item{inp_riveratlas_meta}{path to metadata table for river atlas variables}

\item{in_gaugestats}{data.table of formatted gauging station summary statistics and hydro-environmental attributes.}
}
\value{
data.table of selected predictor variable codes, names, category,
attribute, sources, references, etc. Most sources are from RiverATLAS
technical documentation available at https://www.hydrosheds.org/page/hydroatlas.
}
\description{
Select candidate predictor variables to use in subsequent modelling and create
a table of formatted names.
}
