% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRmapping_functions.R
\name{selecttrain_rf}
\alias{selecttrain_rf}
\title{Select and train final random forest}
\usage{
selecttrain_rf(
  in_rf,
  in_learnerid = NULL,
  in_task = NULL,
  insamp_nfolds = NULL,
  insamp_nevals = NULL
)
}
\arguments{
\item{in_rf}{\link[mlr3]{ResampleResult} of learner to use or \link[mlr3]{BenchmarkResult} from which to extract learner.}

\item{in_task}{\link[mlr3]{Task} containing predictor variables to subset.}

\item{insamp_nfolds}{(optional) number of cross-validation folds to adjust in inner (hyperparameter tuning) cross-validation}

\item{insamp_nevals}{(optional) number of cross-validation repetitions in inner (hyperparameter tuning) cross-validation}

\item{in_lrnid}{id of learner to extract from \link[mlr3]{BenchmarkResult} (e.g., "oversample.classif.ranger")}
}
\value{
list containing the outer resampling (i.e. performance cross-validation) results named ('rf_outer'),
the trained learner ('rf_inner') and the task on which it was trained ('task').
}
\description{
Select final random forest learner and train it, optionally adjusting inner sampling strategy
for hyperparameter tuning (i.e., cross-validation strategy).
}
