% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRmapping_functions.R
\name{set_cvresampling}
\alias{set_cvresampling}
\title{Set cross-validation resampling}
\usage{
set_cvresampling(rsmp_id, in_task, outsamp_nrep, outsamp_nfolds)
}
\arguments{
\item{rsmp_id}{type of resampling, 'cv', 'repeated_cv', or 'repeated-spcv-coords'}

\item{in_task}{\link[mlr3]{Task} to resample}

\item{outsamp_nrep}{number of cross-validation repetitions}

\item{outsamp_nfolds}{number of cross-validation folds.}
}
\value{
a \link[mlr3]{mlr_resamplings_repeated_cv} or \link[mlr3spatiotempcv]{ResamplingRepeatedSpCVCoords}.
}
\description{
Create a \link[mlr3]{Resampling} for model cross-validation.
}
\details{
For more information on the cross-validation strategy used in this study, see
section IV in the Supplementary Information of Messager et al. 2021 at \link{https://www.nature.com/articles/s41586-021-03565-5}.
For general information on resampling in mlr3, see \link{https://mlr3book.mlr-org.com/resampling.html}.
}
