% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRmapping_functions.R
\name{sfformat_wintri}
\alias{sfformat_wintri}
\title{Format to sf and project to Winkel Tripel}
\usage{
sfformat_wintri(in_sp)
}
\arguments{
\item{in_sp}{object to be converted into an object class \code{sf}}
}
\value{
\code{[sf](https://r-spatial.github.io/sf/articles/sf1.html)} object
}
\description{
Convert an object (e.g. data.frame or data.table) to simple feature (spatial object)
and project it to \href{https://en.wikipedia.org/wiki/Winkel_tripel_projection}{Winkel Tripel}.
}
