% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRmapping_functions.R
\name{tabulate_benchmarks}
\alias{tabulate_benchmarks}
\title{Tabulate benchmarks}
\usage{
tabulate_benchmarks(in_bm, in_bmid, inp_resdir = NULL, interthresh = NULL)
}
\arguments{
\item{in_bm}{\link[mlr3]{BenchmarkResult} or file name of one stored on disk.}

\item{in_bmid}{(character)Type of model. one of 'classif1' (selection of classification algorithm),
'regr1' (selection of regression model), 'classif2' (comparison between without and without variable selection).}

\item{inp_resdir}{(character) path to results directory where in_bm is stored.}

\item{interthresh}{(numeric) flow intermittence probability threshold above which
to classify records as non-perennial.}
}
\value{
list of two one-row table, one with hyperparameter values and
the other with performance metrics.
}
\description{
Function to compile hyperparameters and performance statistics for a
given random forest model into two summary tables.
}
\details{
Function used to produce Tables S2 and Tables S3.
}
