% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRmapping_functions.R
\name{tabulate_globalsummary}
\alias{tabulate_globalsummary}
\title{Tabulate global summary}
\usage{
tabulate_globalsummary(
  outp_riveratlaspred,
  inp_riveratlas,
  inp_riveratlas_legends,
  idvars,
  interthresh,
  castvar,
  castvar_num = TRUE,
  weightvar,
  valuevar,
  valuevarsub,
  binfunc = NULL,
  binarg = NULL,
  bintrans = NULL,
  na.rm = T,
  tidy = FALSE,
  nolake = TRUE,
  mincutoff = 0
)
}
\arguments{
\item{outp_riveratlaspred}{(character) full path to .csv table containing flow
intermittence predictions for global river network}

\item{inp_riveratlas}{(character) full path to .csv. table containing attributes
for global river network}

\item{inp_riveratlas_legends}{(character) full path to .xlsx table containing
metadata on river atlas attributes (meaning of category acronyms and codes e.g., for land cover)}

\item{idvars}{(character) name of the variable (usually categorical) whose
categories to use for producing summary statistics of flow prevalence (e.g., climate or country)}

\item{interthresh}{(numeric) flow intermittence probability threshold above which
to classify records as non-perennial.}

\item{castvar}{(character) name of the column that will be used to define bins/table columns (e.g., discharge) —
the equivalent of \code{binvar} in \link{bin_dt}.}

\item{castvar_num}{(boolean) whether casting column (castvar) is a numeric (used for sorting properly).}

\item{weightvar}{(character) variable to weigh proportion by (e.g. river reach length).}

\item{valuevar}{(character) variable to summarize (e.g. intermittency class).}

\item{valuevarsub}{(character) value of \code{valuevar} to summarize (e.g. '1' for intermittent rivers)}

\item{binfunc}{(character) binning approach. One of 'manual', 'equal_length', 'equal_freq'. See \link{bin_dt}.}

\item{binarg}{(numeric) binning argument, depends on binning approach. See \link{bin_dt}.}

\item{bintrans}{(character or numeric) transformation of \code{binvar}, default is NULL.}

\item{na.rm}{(logical) whether to include NAs.}

\item{tidy}{(boolean) whether to keep in tidy long form (TRUE) or to cast into
formatted shape (FALSE; i.e., idvars as rows and castvar as columns + a world row)}

\item{nolake}{(boolean) whether to include sections of river reaches that intersect lakes (FALSE) or not (TRUE)}

\item{mincutoff}{(numeric) smallest river reaches to include in computation, in terms of long-term mean annual flow (m3/s; e.g., 0.1)}
}
\value{
data.table
}
\description{
Create a summary table of the predicted prevalence of IRES across continuous size classes
and categorical classes.
}
\details{
this function is used to produce Table 1 in Main Text, Extended Data Figure 1c, and
Supplementary Data tables from Messager et al. (2021).

bintrans can either be 'log' (for natural log) or a numeric exponent to transform
according to x^bintrans.
}
