% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRmapping_functions.R
\name{test_thresholdsensitivity}
\alias{test_thresholdsensitivity}
\title{Test threshold sensitivity}
\usage{
test_thresholdsensitivity(
  in_gpredsdt,
  in_rivpred,
  threshrange_gauges = seq(0.3, 0.7, 0.01),
  threshrange_network = seq(0.45, 0.55, 0.01),
  mincutoff = 0.1,
  gaugescol = "IRpredprob_CVnosp",
  netcol = "predbasic800"
)
}
\arguments{
\item{in_gpredsdt}{data.table of model predictions for gauging stations. Here, output from \link{bind_gaugepreds}.}

\item{in_rivpred}{data.table of model predictions for global river network. Here, output from \link{netpredformat}.}

\item{threshrange_gauges}{numerical vector of probability threshold values for
which to assess and plot model predictions and performance for gauging stations.}

\item{threshrange_network}{numerical vector of probability threshold values for which to produce predictions
of flow intermittence for the global network (set narrower than \code{thresrange_gauges} as each one takes
much longer to compute).}

\item{mincutoff}{minimum long-term mean annual flow (MAF) to include in assessment of gauges and network
(i.e., mincutoff == 0.1 means that only gauging stations and river reaches with a WaterGAP estimated MAF >= 0.1
are included in the sensitivity analysis)}

\item{gaugescol}{name of the column in \code{in_gpredsdt} containing the predicted probability of flow intermittence.}

\item{netcol}{name of the column in \code{in_rivpred} containing the predicted probability of flow intermittence.}
}
\value{
list with three elements:
\itemize{
\item A ggplot of the probability threshold ranges for gauges >= 10m3/s and gauges
< 10 m3/s that maximize balanced accuracy, bias, raw accuracy, and |sensitivity-specificity|.
\item A ggplot of the predicted global prevalence of IRES as a function of probability thresholds for two gauge size classes (<10 and >=10 m3/s)
\item data.table containing the predicted global prevalence of IRES based on the range of probability threshold \code{threshrange_network}.
}
}
\description{
Test sensitivity of model predictions to flow intermittence probability threshold
used to classify global river reaches into non-perennial and perennial classes.
}
