% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRmapping_functions.R
\name{threshold_dat}
\alias{threshold_dat}
\title{Threshold data (deprecated)}
\usage{
threshold_dat(bmres)
}
\arguments{
\item{bmres}{\link[mlr3]{BenchmarkResult}}
}
\value{

}
\description{
Compute Balanced accuracy (BACC) after classifying the predictions from a
probability ml model into binary classes for every probability threshold value
between 0.4 and 0.6 in 0.01 increments. \cr
}
\details{
Utility function used within \code{\link{formatmisclass_bm}}.
Not currently used in final analysis.
}
