% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRmapping_functions.R
\name{threshold_misclass}
\alias{threshold_misclass}
\title{Threshold misclassification}
\usage{
threshold_misclass(i = 0.5, in_preds)
}
\arguments{
\item{i}{(numeric) Threshold to be used for converting probabilities
predictions to binary predictions, limits=[0,1].}

\item{in_preds}{Either:
\enumerate{
\item a \link[mlr3]{PredictionClassif} or
\item a data.table of predictions for a set of CV repetitions as formatted by
\code{\link{analyze_benchmark}}. The column of the predicted probability of
getting a 'positive' must be called \code{prob.1}.
}}
}
\value{
A single-row data.table with four columns:
\itemize{
\item i - the threshold used to compute the classification statistics
\item misclass - misclassification rate [0-1]
\item sens - sensitivity [0-1]
\item spec - specificity [0-1]
}
}
\description{
Compute the misclassification rate, sensitivity, and specificity of probability
predictions to a binary classification problem from a random forest learner
based on a given threshold.
}
\details{
\itemize{
\item Misclassification rate is the proportion of misclassified
records
\item Sensitivity is the proportion of actual positives (1) that are
correctly identified as such (the proportion of intermittent rivers that
are identified as intermittent)
\item Specificity is the proportion of actual negatives that are correctly
identified as such (the proportion of perennial rivers that are identified as
perennial)
} \cr
See \url{https://en.wikipedia.org/wiki/Sensitivity_and_specificity}
}
\section{Warning}{

This function was only tested for the outputs of a probability random forest
(using a classification framework) and a regression forest (using 0 or 1 as
dummy continuous variables)
}

