% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRmapping_functions.R
\name{weighted_sd}
\alias{weighted_sd}
\title{Weighted standard deviation}
\usage{
weighted_sd(x, w = NULL, na.rm = FALSE)
}
\arguments{
\item{x}{An object containing the values whose weighted mean is to be
computed.}

\item{w}{An integer vector of weights the same length as x giving the
weights to use for elements of x (weights are > 1, not decimals).}

\item{na.rm}{(logical) Whether NA values in x should be stripped before the
computation proceeds.}
}
\value{
A length-one numeric vector.
}
\description{
Computes a weighted standard deviation
}
\examples{
```{r}
wt <- c(5, 4, 3 ,2 , 1)
x <- c(1, 2, 3, 4, 5)
xm <- weighted.mean(x, wt)
xsd <- weighted_sd(x, wt)
```

}
