% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRmapping_functions.R
\name{weighted_vimportance_nestedrf}
\alias{weighted_vimportance_nestedrf}
\title{Weighted mean of variable importance for resampled RF learner}
\usage{
weighted_vimportance_nestedrf(rfresamp, pvalue = TRUE, pvalue_permutn = 100)
}
\arguments{
\item{rfresamp}{the \link[mlr3]{ResampleResult} from a classification RF
of type \link[mlr3learners]{mlr_learners_classif.ranger}
or \link[mlr3learners.partykit]{mlr_learners_classif.cforest}}

\item{pvalue}{(logical) Whether to compute p-values.}

\item{pvalue_permutn}{(integer) number of permutations to use in p-value
calculations}
}
\value{
A data.table with, when all logicals are TRUE, the following columns.
\describe{
\item{varnames} - predictor variable name
\item{imp_wmean} - weighted mean of variable importance across resampling
instances
\item{imp_wsd} - weighted standard deviation of variable importance across
resampling instances
\item{pvalue_wmean} -weighted mean of variable p-value across resampling
instances
}
}
\description{
Compute mean and standard deviation of variable importance and mean of
p-value across resampling instances (e.g. folds and repetitions) weighted by
resampling prediction accuracy.
}
\details{
See \code{\link{extract_impperf_nestedrf}} for more details on computations.
}
\section{Warning}{

Does not accept error measure for weighting
\itemize{
\item could be amended to be based on 1-error
}
}

\section{documentation to-do}{

Can add an example down the line, add source.
}

\seealso{
\code{\link{extract_impperf_nestedrf}} and \code{\link{ggvimp}} and
\code{\link{benchmark_featsel}}
}
