% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRmapping_functions.R
\name{write_gaugepreds}
\alias{write_gaugepreds}
\title{Write gauge predictions}
\usage{
write_gaugepreds(in_gaugep, in_gpredsdt, outp_gaugep)
}
\arguments{
\item{in_gaugep}{\link[sf]{sf} points and hydro-environmental attributes for all gauging stations.}

\item{in_gpredsdt}{final formatted model predictions for all gauging stations. Output from \link{bind_gaugepreds}.}

\item{outp_gaugep}{full path to which the geopackage will be written, including the file name and extension.}
}
\value{
path to output geopackage i.e. \code{outp_gaugep}
}
\description{
Join model predictions for gauging stations to point feature class of gauging stations,
then write the result of the merging operation to disk as a GeoPackage.
}
\details{
for more details on GeoPackage (extension .gpkg), see \link{https://www.geopackage.org/}.
}
