% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRmapping_functions.R
\name{write_netpreds}
\alias{write_netpreds}
\title{Write river network predictions}
\usage{
write_netpreds(
  in_network,
  in_rftuned,
  in_predvars,
  predcol,
  discharge_interval = c(-Inf, Inf),
  interthresh = 0.5,
  outp_riveratlaspred
)
}
\arguments{
\item{in_network}{data.table of formatted global river network with all predictor
hydro-environmental variables. Output from \link{rformat_network}.}

\item{in_rftuned}{trained random forest model. Output from \link{selecttrain_rf};
list containing inner resampling results.}

\item{in_predvars}{data.table of predictor variable codes, names and attributes.
Output from \link{selectformat_predvars}.}

\item{predcol}{(character) name of the column to which the predicted probability
of flow intermittence for each reach be written.}

\item{discharge_interval}{(vector of two numeric values) discharge criterion
to subset global river network. Range of long-term mean annual discharge values
for which to produce model predictions.}

\item{interthresh}{(numeric) between 0 and 1 (inclusive), probability threshold
above which to classify river and stream reaches as non-perennial (e.g., 0.50).}

\item{outp_riveratlaspred}{(character) ull path to which the output csv file will be written,
including the file name and extension.}
}
\value{
path to output csv file.
}
\description{
Make model predictions for global river network and write out csv table.
}
\details{
the date is added to the output file name in the format
\code{outp_riveratlaspred_YYYYMMDD}.
}
