% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRmapping_functions.R
\name{comp_GRDCdurfreq}
\alias{comp_GRDCdurfreq}
\title{Compute intermittency statistics for GRDC gauging stations}
\usage{
comp_GRDCdurfreq(
  path,
  in_gaugep,
  maxgap,
  mdurthresh = 1,
  windowsize = 100,
  fullwindow = FALSE,
  monthsel = NULL,
  verbose = FALSE
)
}
\arguments{
\item{path}{file path to a GRDC-formatted streamflow time series table}

\item{maxgap}{maximum number of days with missing data beyond which a year is
not used in the computation of statistics}

\item{mdurthresh}{threshold of mean annual number of zero-flow days beyond
which to classify gauge as intermittent.}

\item{monthsel}{selected months to compute the statistics over}
}
\value{
One row data.table with the following columns: \cr
\describe{
\item{GRDC_NO} - (char) unique identifier for the gauge
\item{firstYear} - (num) first year on full record
\item{lastYear} - (num) last year on full record
\item{totalYears} - (int) total number of years on full record
\item{firstYear_kept} - (num) first year on record with < maxgap missing days
\item{lastYear_kept} - (num) first year on record with < maxgap missing days
\item{totalYears_kept} - (int) total number of years with < maxgap missing days
\item{totaldays} - (num) total number of days with discharge data
\item{sumDur} - (int) total number of days with discharge = 0
\item{mDur} - (num) mean number of days/year with discharge = 0
\item{mFreq} - (num) mean number of periods with discharge = 0
(with at least one day of flow between periods)
}
}
\description{
Determine general characteristics of the whole time series and of the subset
of years that have less than a given threshold of missing data as well as
intermittency statistics. The intermittency statistics can be computed for a
subset of months of the year (e.g. only winter months)
}
