% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRmapping_functions.R
\name{format_gaugestats}
\alias{format_gaugestats}
\title{Format gauge statistics}
\usage{
format_gaugestats(in_gaugestats, in_gaugep, yearthresh)
}
\arguments{
\item{path}{file path to a GRDC-formatted streamflow time series table}

\item{maxgap}{maximum number of days with missing data beyond which a year is
not used in the computation of statistics}

\item{monthsel}{selected months to compute the statistics over}
}
\value{
One row data.table with the following columns: \cr
\describe{
\item{GRDC_NO} - (char) unique identifier for the gauge
\item{firstYear} - (num) first year on full record
}
}
\description{
Format gauge attributes
}
