% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRmapping_functions.R
\name{qc_onde}
\alias{qc_onde}
\title{HydroSHEDS Drainage area (DA, km2)',}
\usage{
qc_onde(
  inp_ondedatdir,
  inp_onderesdir,
  inp_riveratlas,
  in_rivpred,
  predcol,
  interthresh = 0.5,
  mincutoff = 0
)
}
\description{
HydroSHEDS Drainage area (DA, km2)',
}
