% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRmapping_functions.R
\name{read_gaugep}
\alias{read_gaugep}
\title{Read gauge points}
\usage{
read_gaugep(in_filestructure, dist)
}
\arguments{
\item{in_filestructure}{named list containing path to point data for gauging
stations, named \code{in_gaugep}}

\item{dist}{maximum distance from the river network beyond which gauging
stations are excluded.}
}
\value{
object of class sf
}
\description{
Import streamgauging station spatial data and attributes, only keeping those
whose point representation is within a given distance from the river network.
}
\details{
The distance from the river network must have been determined
beforehand and be an attribute of the gauge stations point data called
\code{station_river_distance}
}
