% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRmapping_functions.R
\name{read_gaugep}
\alias{read_gaugep}
\title{Read gauge points}
\usage{
read_gaugep(
  inp_GRDCgaugep,
  inp_GSIMgaugep,
  inp_riveratlas2,
  in_monthlydischarge = NULL
)
}
\arguments{
\item{inp_GRDCgaugep}{path to point data for GRDC gauging stations.}

\item{inp_GSIMgaugep}{path to point data for GSIM gauging stations.}

\item{inp_riveratlas2}{path to attribute table of RiverATLAS v1.0.9}

\item{in_monthlydischarge}{data frame of naturalized monthly discharge for
every river reach in HydroSHEDS}
}
\value{
object of class \link[sf]{sf}
}
\description{
Import streamgauging station spatial data and attributes for GRDC and GSIM.
Add new and updated environmental predictors from RiverATLAS v1.0.9 and, optionally,
estimates of monthly discharge from WaterGAP v2.2.
}
